refactor: split grid.py and move all bbox related functions to util.bbox
