+++ mapproxy/test/unit/test_client.py
@@ -22,7 +22,7 @@ import pytest
 from mapproxy.client.http import HTTPClient, HTTPClientError
 from mapproxy.client.tile import TileClient, TileURLTemplate
 from mapproxy.client.wms import WMSClient, WMSInfoClient
-from mapproxy.grid import tile_grid
+from mapproxy.grid.tile_grid import tile_grid
 from mapproxy.layer import MapQuery, InfoQuery
 from mapproxy.request.wms import (
     WMS111MapRequest,
